Move config file under `configs` directory
