+++ src/utils/redisCache.util.js
@@ -1,5 +1,5 @@
 const redisClient = require("../client/redis.client");
-const { REDIS_TTL } = require("../config");
+const { REDIS_TTL } = require("../configs/env.config");
 
 const ttl = REDIS_TTL || 3600;
